Enrollment: organize components in subfolders (invitation, request, list), fix related imports
