+++ components/Enrollment/Request/EnrollmentRequestCreate.tsx
@@ -8,35 +8,33 @@ import { notifications } from '@mantine/notifications'
 import { Projects, ProjectsQueryKey } from '@/services/projects'
 import sanitizeHtml from 'sanitize-html'
 
-import Enrollment from '@/entities/Enrollment/Enrollment'
 import { EnrollmentRequestInput } from '@/entities/Enrollment/EnrollmentRequestInput'
-import TextEditor from '../Common/TextEditor/TextEditor'
+import TextEditor from '@/components/Common/TextEditor/TextEditor'
 
-interface EnrollmentRequestProps {
+interface EnrollmentRequestCreateProps {
   projectId: number
-  enrollment: Enrollment
 }
 
-export const EnrollmentRevoke = (props: EnrollmentRequestProps): React.JSX.Element => {
+export const EnrollmentRequestCreate = (props: EnrollmentRequestCreateProps): React.JSX.Element => {
   const form = useForm({ initialValues: { message: '' } })
   const queryClient = useQueryClient()
 
-  const revokeEnrollmentMutation = useMutation({
-    mutationFn: async (adminOptions: EnrollmentRequestInput) =>
-      Projects.revokeEnrollment(props.projectId, props.enrollment.user.id, adminOptions),
+  const enrollmentRequestMutation = useMutation({
+    mutationFn: async (enrollmentRequest: EnrollmentRequestInput) => {
+      return Projects.requestEnrollment(props.projectId, enrollmentRequest)
+    },
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: [ProjectsQueryKey, props.projectId] })
+      queryClient.invalidateQueries({ queryKey: [ProjectsQueryKey] })
       notifications.show({
-        title: 'Inscripción revocada',
-        message: 'El usuario ya no es miembro del proyecto',
-        color: 'blue',
+        title: 'Solicitud de inscripción enviada',
+        message: 'Espera a que el líder del proyecto acepte tu solicitud',
       })
       modals.closeAll()
     },
     onError: (error) => {
-      console.error(error)
+      console.error('Enrollment request failed:', error)
       notifications.show({
-        title: 'Error al revocar la inscripción',
+        title: 'Error al enviar la solicitud de inscripción',
         message: 'Por favor, inténtalo de nuevo más tarde',
         color: 'red',
       })
@@ -45,25 +43,21 @@ export const EnrollmentRevoke = (props: EnrollmentRequestProps): React.JSX.Eleme
 
   const handleSubmit = (values: typeof form.values) => {
     const sanitizedMessage = sanitizeHtml(values.message)
-    revokeEnrollmentMutation.mutate({
-      message: sanitizedMessage,
-    })
+    enrollmentRequestMutation.mutate({ message: sanitizedMessage })
   }
 
   return (
     <form onSubmit={form.onSubmit(handleSubmit)}>
       <LoadingOverlay
-        visible={revokeEnrollmentMutation.isPending}
+        visible={enrollmentRequestMutation.isPending}
         zIndex={1000}
         overlayProps={{ radius: 'sm', blur: 2 }}
       />
-      <Text style={{ marginBottom: '1rem' }}>
-        [Opcional] Puedes enviar un mensaje al usuario para explicar la decisión de revocar la
-        inscripción
-      </Text>
+      <Text size="lg">[Opcional] Envía un mensaje al líder o admin del proyecto</Text>
+      <Text size="sm">Puedes indicar tus motivaciones, habilidades, disponibilidad, etc.</Text>
       <TextEditor onChange={(content) => form.setValues({ message: content })} />
-      <Button type="submit" fullWidth mt="md" aria-label="Revocar inscripción" color="red">
-        Revocar inscripción
+      <Button type="submit" fullWidth mt="md">
+        Solicitar
       </Button>
     </form>
   )
