+++ components/Project/ProjectDetails.tsx
@@ -37,9 +37,9 @@ import { CurrentUserQueryOptions } from '@/services/currentUser'
 import { EnrollmentRequestsQueryKey, Projects, ProjectsQueryKey } from '@/services/projects'
 import { EnrollmentRequestShow } from '@/entities/Enrollment/EnrollmentRequestShow'
 
-import EnrollmentButtonRequest from '@/components/Enrollment/EnrollmentButtonRequest'
-import { EnrollmentList } from '../Enrollment/EnrollmentList'
-import { EnrollmentRequestAdmin } from '../Enrollment/EnrollmentRequestAdmin'
+import EnrollmentButtonRequest from '@/components/Enrollment/Request/EnrollmentButtonRequest'
+import { EnrollmentList } from '@/components/Enrollment/List/EnrollmentList'
+import { EnrollmentRequestAdmin } from '@/components/Enrollment/Request/EnrollmentRequestAdmin'
 import { NotLoggedError } from '@/components/Account/NotLoggedError'
 import ProjectPDF from '@/components/Project/ProjectPDF'
 import SkeletonFull from '@/components/Common/Loader/SkeletonFull'
